Carry money as BigDecimal, name the actualisation methods, spell out FC
Money no longer round-trips through binary floating point. Every amount in the
lifecare service model and the financial assistance API models is a BigDecimal,
matching the entities, which already were. Double survives only where the
generated FamilyCare DTOs demand it, and MapperUtil.toAmount/toWireAmount are
the single boundary that converts. The OpenAPI contract only loses its
"format: double" hints - the JSON stays a number, so clients are unaffected.

ActualisationService.create/list become createActualisation/listActualisations.

ExpenseTypeMapper's reverse map is written out by hand instead of derived from
the forward one, so both directions read the same way. Because two hand-written
maps can drift - and these map money read back off a previous calculation - a
test asserts they stay exact inverses, keyed by the normalized name. Removing
one entry fails the build.

Also finishes the FC sweep: identifiers that carried Fc mid-word
(FC_NAME_BY_COST_TYPE, costTypeForFcName, SsbtekToFcIncomeMapper, several test
names) now spell out FamilyCare, and the javadoc that the EB and FC expansions
left fragmented is re-flowed - the formatter wraps overlong lines but never
re-joins them, so successive widenings had shredded whole paragraphs.

+++ src/main/java/se/sundsvall/caremanagement/lifecare/service/model/CalculationExpenseView.java
@@ -1,11 +1,13 @@
 package se.sundsvall.caremanagement.lifecare.service.model;
 
+import java.math.BigDecimal;
+
 /**
  * A single expense row on a Lifecare normberäkning, as read for display — used for both the regular (UTGIFTER) and the
  * special-expense (LEVNADSKOSTNADER I ÖVRIGT) arrays.
  */
 public record CalculationExpenseView(
 	String type,
-	Double appliedAmount,
-	Double approvedAmount) {
+	BigDecimal appliedAmount,
+	BigDecimal approvedAmount) {
 }
